Small bug fixes

+++ include/hash.h
@@ -39,7 +39,7 @@ void initializeTable(void);
 int isThreeRep(const Repetition* r, const uint64_t hash);
 uint64_t hashPosition(const Board* b);
 uint64_t makeMoveHash(uint64_t prev, Board* b, const Move m, const History h);
-int changeTurn(const uint64_t prev);
+uint64_t changeTurn(const uint64_t prev);
 
 inline int isRepetition(const Repetition* rep, const uint64_t hash) {return rep->hashTable[rep->index - 4] == hash;}
 inline void addHash(Repetition* rep, uint64_t hash) {rep->hashTable[rep->index++] = hash;}
